+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v4.5.0
 * Fixed: If you had access an element but none of the extended data columns you could still match on the extended data row using a query with no filters. This would result in the row being counted in `totalHits`, but it could not be retrieved from Accumulo. NOTE: This change requires that all extended data values are re-indexed in ElasticSearch.
+* Fixed: If the visibility of an element changes, Vertexium will now re-index all associated extended data to ensure that they can be found according to the new visibility.
 
 # v4.4.2
 * Fixed: `EdgeInfo.getDirection` was reversed when loaded from InMemoryVertex.